# Conflicts:
#	.idea/misc.xml
#	gradle.properties
#	src/com/ansorgit/plugins/bash/lang/lexer/_BashLexerBase.java
#	src/com/ansorgit/plugins/bash/lang/parser/util/ParserUtil.java
#	src/com/ansorgit/plugins/bash/runner/repl/BashConsoleRunner.java

+++ src/com/ansorgit/plugins/bash/lang/parser/command/CommandParsingUtil.java
@@ -1,13 +1,10 @@
 /*
- * Copyright 2013 Joachim Ansorg, mail@ansorg-it.com
- * File: CommandParsingUtil.java, Class: CommandParsingUtil
- * Last modified: 2013-04-30
+ * Copyright (c) Joachim Ansorg, mail@ansorg-it.com
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
+ * http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -30,14 +27,15 @@ import com.intellij.psi.tree.TokenSet;
 
 /**
  * Parsing function for commands.
- * <p/>
- * User: jansorg
- * Date: Nov 26, 2009
- * Time: 9:16:36 PM
+ * <br>
+ * @author jansorg
  */
 public class CommandParsingUtil implements BashTokenTypes, BashElementTypes {
     private final static TokenSet assignmentSeparators = TokenSet.create(LINE_FEED, SEMI, WHITESPACE);
     private final static TokenSet validWordTokens = TokenSet.create(ARITH_NUMBER);
+
+    private CommandParsingUtil() {
+    }
 
     public static boolean readCommandParams(final BashPsiBuilder builder) {
         return readCommandParams(builder, TokenSet.EMPTY);
@@ -69,30 +67,12 @@ public class CommandParsingUtil implements BashTokenTypes, BashElementTypes {
         return ok;
     }
 
-    public enum Mode {
-        /**
-         * Only accept an ASSIGNENT_WORD or ARRAY_ASSIGNMENT_WORD in front .
-         * The =value part is mandatory.
-         */
-        StrictAssignmentMode,
-        /**
-         * Optional =value part and allows simple word tokens and variable names in front
-         */
-        LaxAssignmentMode,
-        /**
-         * Only single word tokens are valid, used for read commands.
-         */
-        SimpleMode
-    }
-
-    private CommandParsingUtil() {
-    }
-
     public static boolean isAssignment(final BashPsiBuilder builder, Mode mode, boolean acceptArrayVars) {
         final IElementType tokenType = builder.getTokenType();
         switch (mode) {
             case SimpleMode:
-                return ParserUtil.isWordToken(tokenType)
+                return (acceptArrayVars && ParserUtil.hasNextTokens(builder, false, ASSIGNMENT_WORD, LEFT_SQUARE))
+                        || ParserUtil.isWordToken(tokenType)
                         || Parsing.word.isWordToken(builder)
                         || Parsing.var.isValid(builder);
 
@@ -148,7 +128,6 @@ public class CommandParsingUtil implements BashTokenTypes, BashElementTypes {
         return ok;
     }
 
-
     /**
      * Reads a single assignment
      *
@@ -163,6 +142,9 @@ public class CommandParsingUtil implements BashTokenTypes, BashElementTypes {
 
         switch (mode) {
             case SimpleMode:
+                if (acceptArrayVars && builder.getTokenType() == ASSIGNMENT_WORD) {
+                    break;
+                }
                 if (!Parsing.word.parseWord(builder)) {
                     assignment.drop();
                     return false;
@@ -280,7 +262,7 @@ public class CommandParsingUtil implements BashTokenTypes, BashElementTypes {
      * Parses an assignment list like a=(1 2 3)
      * Grammar (selfmade):
      * assignment_list ::= "(" array_assignment {" " {array_assignment}* ")"
-     * <p/>
+     * <br>
      * array_assignment ::=
      * WORD_EXPR
      * |    STRING
@@ -350,6 +332,22 @@ public class CommandParsingUtil implements BashTokenTypes, BashElementTypes {
 
         marker.done(VAR_ASSIGNMENT_LIST);
         return true;
+    }
+
+    public enum Mode {
+        /**
+         * Only accept an ASSIGNENT_WORD or ARRAY_ASSIGNMENT_WORD in front .
+         * The =value part is mandatory.
+         */
+        StrictAssignmentMode,
+        /**
+         * Optional =value part and allows simple word tokens and variable names in front
+         */
+        LaxAssignmentMode,
+        /**
+         * Only single word tokens are valid, used for read commands.
+         */
+        SimpleMode
     }
 
 }
